+++ Api/Contracts/SupplyChain/ItemResponseDTO.cs
@@ -6,14 +6,14 @@ public class ItemResponse
     public int UomId { get; set; }
     public int CategoryId { get; set; }
     public string ItemName { get; set; } = string.Empty;
-    public int MinStockLevel { get; set; }
-    public int MaxStockLevel { get; set; }
+    public decimal MinStockLevel { get; set; }
+    public decimal MaxStockLevel { get; set; }
     public bool IsActive { get; set; }
     public string UomName { get; set; } = string.Empty;
     public string CategoryName { get; set; } = string.Empty;
-    public int CurrentStock { get; set; }
+    public decimal CurrentStock { get; set; }
     public double StockPercentage => MinStockLevel > 0
-        ? (double)CurrentStock / MinStockLevel * 100
+        ? (double)(CurrentStock / MinStockLevel) * 100
         : 100;
     public bool IsLowStock => CurrentStock < MinStockLevel;
 }
